moved check_genesis_block function to p2pool.bitcoin.helper
